Add BuildArgs and refactor build structure
