+++ mlc_llm/__init__.py
@@ -3,4 +3,5 @@ from . import quantization
 from . import relax_model
 from . import transform
 from . import utils
-from .build import build_model
+from . import core
+from .core import build_model, BuildArgs
